refactor: extract repo routes from router into routes/repos.ts (#153 follow-up)
Move handleListRepos, handleUpdateRepoMetadata, handleGetRepoMetadata
and their supporting constants/types/helpers into a dedicated route
module, matching the pattern established by secrets, model-preferences,
and integration-settings.
